Change interaction to be specifically with the original element
Change the interaction model to be explicitly with the original ul
rather than with the generated div.containerobj. This will require us to
proxy all interaction with the actual widget using setter/getter
methods, but enables us to cleanly .destroy() the widget and forces
consistent interaction. As a happy coincidence it also documents the ID
attribute values used for the widget, and allows the user to configure
it within the bounds of its naming rules.